Moved Game class location
