+++ src/main/java/com/arthur/blackjack/Main.java
@@ -1,5 +1,7 @@
 package com.arthur.blackjack;
 
+import com.arthur.blackjack.core.Game;
+
 public class Main {
     public static void main(String[] args) {
         Game game = new Game();
