+++ src/main/java/com/arthur/blackjack/core/Game.java
@@ -1,4 +1,4 @@
-package com.arthur.blackjack;
+package com.arthur.blackjack.core;
 
 import com.arthur.blackjack.component.Deck;
 import com.arthur.blackjack.player.Dealer;
